Change UnityTheme tristate CheckButton to LibConfig RadioButtonGroup
Also change useUnityTheme=0 to Never and useUnityTheme=2 to Always. I
am not sure why the previous behavior was written this way. I think
there was a bug in my tristate serialization code.

KDE's HIG mention to avoid tristate buttons so it's changed to a
RadioButton group.

Begins migration to QQC2.